feat(cli): create output folder if it's not present. (#280)
Creates the destination/output folder if it is not preset and the
`--dest-create` flag is provided in the command.

I decided to make it a command instead of just creating the
folder just incase the provided destination path was a mistake or
a typo and you don't want your fonts to be exported there.